+++ src/vid_gl2/GLEW/glew.h
@@ -222,11 +222,11 @@ extern "C" {
 #define GL_VERSION_1_1 1
 
 #if defined(__APPLE__)
-typedef unsigned long GLenum;
-typedef unsigned long GLbitfield;
-typedef unsigned long GLuint;
-typedef long GLint;
-typedef long GLsizei;
+typedef unsigned int GLenum;
+typedef unsigned int GLbitfield;
+typedef unsigned int GLuint;
+typedef int GLint;
+typedef int GLsizei;
 #else
 typedef unsigned int GLenum;
 typedef unsigned int GLbitfield;
